TransferFile can now write to supplied streams.
Added an additional method signature for TransferFile that accepts a
stream. The filename version of TransferFile now just wraps the stream
version. BinaryWriter is no longer used internally, as it was
unnecessary for writing a byte array, and closed the stream when in a
using block.

This change is backward-compatible.